+++ src/providers/flow/tokenizer/arabic.ts
@@ -1,0 +1,5 @@
+import { DefaultTokenizer } from './default'
+
+export class ArabicTokenizer extends DefaultTokenizer {
+  protected readonly wordSeparatorPattern: RegExp = /[\[\]()<>"'.,|; `!?،؛]/
+}