+++ src/providers/provider.ts
@@ -2,22 +2,31 @@ export abstract class Provider {
   abstract readonly category: string
   abstract completions: Array<string>
 
-  static readonly wordSeparatorRegex = /(\.|,|;|:|'|"|!|\?|-|\)|\]|\}|\/| |Enter)/
+  static readonly wordSeparatorRegex = /(\.|,|;|:|'|"|!|\?|-|\)|\]|\}|\/| |Enter)/g
   static readonly placeholder: string = '#{}'
 
   matchWith(input: string): Completion[] {
-    // TODO: Improve filtering with weights
-    const suggestions = this.completions.filter(val => val.includes(input))
-      .map(sugg => {
-        return {category: this.category, value: sugg}
+    const inputLowered = input.toLowerCase()
+    const inputHasUpperCase = /[A-Z]/.test(input)
+
+    // case-sensitive logic if input has an upper case.
+    // Otherwise, uses case-insensitive logic
+    const suggestions = this.completions
+      .filter((suggestion) =>
+        inputHasUpperCase
+          ? suggestion.includes(input)
+          : suggestion.toLowerCase().includes(inputLowered)
+      )
+      .sort((a,b) => a.localeCompare(b))
+      .map((suggestion) => {
+        return { category: this.category, value: suggestion }
       })
-      .sort((a, b) => a.value.length - b.value.length)
 
     return suggestions
   }
 }
 
 export interface Completion {
-  category: string,
+  category: string
   value: string
 }